Median case 2 build: fully functional
(Hopefully) fully functional build, as it does work for both median 2 case and base case (build hard-coded to execute median 2 file atm). Fixed memory errors and non-generalized parts in previous build. To do next: make it so it gets a valid combination and immediately jumps to the state from that combination, and trim/change data structures as to be less memory intensive (use ints instead of bool arrays and use bitmask-based graph instead of array-based graph).