Fully functional non-optimised build
Fully functional build. I have also made it so that it no longer has to save all combinations into an array and instead it verifies if a single combination is valid and if it is, instantly adds it to heap and stack and keeps going to the next combination. To do: still change graph from arrays to bitmasks and boolean arrays into ints, and add function to calculate minimum vertex cover so the combinations function can calculate all combinations of passengers from size minimum vertex cover to size boats capacity.