Compatibility sets build
Not yet functional for median 1. Added function that returns a Vector** with each element being a Vector* containing the elements that are compatible between them (it essentially separates incompatible psgs in different sets with the ones that are compatible with them). To do: perhaps add a function that, for every state, for every compatibility set (each Vector* from Vector**), checks how many passengers are in the current side. Then, it subtracts x = amntPsgsInSide - amntPsgsFromSetInSide and if x <= currentBoatsCapacity, it crosses the passengers in the current side that are NOT in the set. I would have to update crossVoid aswell, adding a special case if the passengers in the current side are all a set, in which case cross void. With this hypothetical implementation (the one from the to do), I imagine it would be wise to save the passengers from each side in each state, and it should no longer require to make combinations.